Move Body[Inserter|Extractor] to web.reactive.function
