+++ spring-web-reactive/src/main/java/org/springframework/web/reactive/function/client/DefaultClientResponse.java
@@ -32,9 +32,9 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.client.reactive.ClientHttpResponse;
-import org.springframework.http.codec.BodyExtractor;
-import org.springframework.http.codec.BodyExtractors;
 import org.springframework.http.codec.HttpMessageReader;
+import org.springframework.web.reactive.function.BodyExtractor;
+import org.springframework.web.reactive.function.BodyExtractors;
 
 /**
  * Default implementation of {@link ClientResponse}.
