+++ spring-web-reactive/src/main/java/org/springframework/web/reactive/function/server/DefaultServerRequest.java
@@ -34,11 +34,11 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpRange;
 import org.springframework.http.MediaType;
-import org.springframework.http.codec.BodyExtractor;
-import org.springframework.http.codec.BodyExtractors;
 import org.springframework.http.codec.HttpMessageReader;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.util.Assert;
+import org.springframework.web.reactive.function.BodyExtractor;
+import org.springframework.web.reactive.function.BodyExtractors;
 import org.springframework.web.server.ServerWebExchange;
 
 /**
