+++ spring-web-reactive/src/test/java/org/springframework/web/reactive/function/client/DefaultClientResponseTests.java
@@ -46,7 +46,7 @@ import org.springframework.http.codec.HttpMessageReader;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.springframework.http.codec.BodyExtractors.toMono;
+import static org.springframework.web.reactive.function.BodyExtractors.toMono;
 
 /**
  * @author Arjen Poutsma
