+++ spring-web-reactive/src/test/java/org/springframework/web/reactive/function/server/DefaultServerRequestTests.java
@@ -52,7 +52,7 @@ import org.springframework.web.server.ServerWebExchange;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.springframework.http.codec.BodyExtractors.toMono;
+import static org.springframework.web.reactive.function.BodyExtractors.toMono;
 
 /**
  * @author Arjen Poutsma
