+++ spring-web-reactive/src/test/java/org/springframework/web/reactive/function/server/RouterFunctionTests.java
@@ -21,7 +21,7 @@ import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 import static org.junit.Assert.assertNotNull;
-import static org.springframework.http.codec.BodyInserters.fromObject;
+import static org.springframework.web.reactive.function.BodyInserters.fromObject;
 
 /**
  * @author Arjen Poutsma
